refactor: ♻️ Isolated parsing of Payment request
