+++ src/components/ui/PaymentRequestTable/PaymentRequestTable.tsx
@@ -1,5 +1,4 @@
 import classNames from 'classnames';
-import { LocalPaymentRequest } from '../../../types';
 import Pager from '../Pager/Pager';
 import PaymentRequestRow from '../PaymentRequestRow/PaymentRequestRow';
